patrimonio-autopilot/requirements.txt
pyyaml
fastapi
sqlalchemy
uvicorn
pyjwt
